refactor(react-ds-global): consolidate tooltip on withTooltip, delete TooltipArea
Move the tooltip engine (useDisclosure hover mode, useIsMounted client-only
portal, arrow positioning, target-span wrapper) out of the standalone
TooltipArea component and inline it into withTooltip, which becomes the single
public tooltip primitive. The HOC now owns the engine and renders the wrapped
component as the tooltip target.

- Relocate TooltipAreaProps into the Tooltip root types.ts as TooltipEngineProps
  (internal) with WithTooltipOptions = Omit<..., children|Message> unchanged.
- Fold the .tooltip-area wrapper styles into the Tooltip styles.css.
- Migrate all TooltipArea stories into withTooltip.stories.tsx via a Demo helper
  that wraps a trigger with withTooltip and preserves the force-open-in-story /
  hover-in-docs behaviour, the stage + trackViewMode decorators, and every
  variant (icon, multiline, maxWidth, lateral, autofit, playground, changeable,
  placements, inline).
- Port all TooltipArea tests to withTooltip.tests.tsx / .ssr / .hydration,
  preserving every assertion including the SSR no-message and hydration
  no-onRecoverableError guarantees; repoint Tooltip.ssr.tests.tsx off TooltipArea
  onto the presentational bubble.
- Delete common/TooltipArea and common/index.ts; drop the common re-export from
  index.ts. Update withTooltip.mdx and stray comments to drop TooltipArea.

Zero TooltipArea references remain in the package.
