+++ packages/react/ds-global/src/lib/component/Tooltip/Tooltip.ssr.tests.tsx
@@ -1,27 +1,20 @@
 import { renderToString } from "react-dom/server";
 import { describe, expect, it } from "vitest";
-import { TooltipArea } from "./common/TooltipArea/index.js";
+import Tooltip from "./Tooltip.js";
 
-describe("TooltipArea SSR", () => {
-  it("renders the trigger on the server without throwing", () => {
-    const html = renderToString(
-      <TooltipArea Message="Tooltip message">Trigger</TooltipArea>,
-    );
+describe("Tooltip SSR", () => {
+  it("renders the message bubble on the server without throwing", () => {
+    const html = renderToString(<Tooltip isOpen>Tooltip message</Tooltip>);
 
-    expect(html).toContain("ds tooltip-area");
-    expect(html).toContain("Trigger");
+    expect(html).toContain("ds tooltip");
+    expect(html).toContain("Tooltip message");
   });
 
-  it("defers the tooltip message to the client (no SSR message markup)", () => {
-    const html = renderToString(
-      <TooltipArea Message="Tooltip message">Trigger</TooltipArea>,
-    );
+  it("carries the open/closed aria state into the server markup", () => {
+    const open = renderToString(<Tooltip isOpen>Open message</Tooltip>);
+    expect(open).toContain('aria-hidden="false"');
 
-    // The message is portalled and mounted client-only, so the server and the
-    // first client render agree (nothing at the call site) — no hydration
-    // mismatch, no forced re-mount. The trigger still carries its stable
-    // aria-describedby, and the message appears once the tooltip mounts.
-    expect(html).not.toContain("Tooltip message");
-    expect(html).toContain("aria-describedby");
+    const closed = renderToString(<Tooltip>Closed message</Tooltip>);
+    expect(closed).toContain('aria-hidden="true"');
   });
 });
