--- packages/react/ds-global/src/lib/component/Tooltip/common/TooltipArea/index.ts
@@ -1,2 +1,0 @@
-export { default as TooltipArea } from "./TooltipArea.js";
-export type * from "./types.js";
