--- packages/react/ds-global/src/lib/component/Tooltip/common/TooltipArea/styles.css
@@ -1,13 +1,0 @@
-.ds.tooltip-area {
-  position: relative;
-  & .target {
-    cursor: pointer;
-    position: relative;
-    display: inline-block;
-
-    & > * {
-      /* biome-ignore lint/complexity/noImportantStyles: The tooltip targets must have no margin in order for tooltip placements to work correctly */
-      margin: 0 !important;
-    }
-  }
-}
