--- packages/react/ds-global/src/lib/component/Tooltip/common/index.ts
@@ -1,1 +1,0 @@
-export * from "./TooltipArea/index.js";
