+++ packages/react/ds-global/src/lib/component/Tooltip/index.ts
@@ -1,4 +1,3 @@
-export * from "./common/index.js";
 export { default as Tooltip } from "./Tooltip.js";
 export type * from "./types.js";
 export { default as withTooltip } from "./withTooltip.js";
