+++ packages/react/ds-global/src/lib/component/Tooltip/withTooltip.tests.tsx
@@ -1,37 +1,44 @@
 import { fireEvent, render, screen, waitFor } from "@testing-library/react";
 import { createPortal } from "react-dom";
 import { describe, expect, it, vi } from "vitest";
-import Tooltip from "./TooltipArea.js";
+import type { WithTooltipOptions } from "./index.js";
+import { withTooltip } from "./index.js";
 
 vi.mock("react-dom", () => ({
   createPortal: vi.fn((children) => children),
 }));
 
-describe("TooltipArea component", () => {
+describe("withTooltip", () => {
   const Message = "Tooltip Message";
-  const Children = <span>Target Element</span>;
+  const Trigger = () => <span>Target Element</span>;
+
+  /** Wrap the trigger with a tooltip and render it. */
+  const renderTooltip = (options?: WithTooltipOptions) => {
+    const Tooltipped = withTooltip(Trigger, Message, options);
+    return render(<Tooltipped />);
+  };
 
   it("renders children", () => {
-    render(<Tooltip Message={Message}>{Children}</Tooltip>);
+    renderTooltip();
     expect(screen.getByText("Target Element")).toBeInTheDocument();
   });
 
   it("renders tooltip message on hover", async () => {
-    render(<Tooltip Message={Message}>{Children}</Tooltip>);
+    renderTooltip();
     const target = screen.getByText("Target Element");
     fireEvent.pointerEnter(target);
     expect(await screen.findByText(Message)).toBeInTheDocument();
   });
 
   it("renders tooltip message on focus", async () => {
-    render(<Tooltip Message={Message}>{Children}</Tooltip>);
+    renderTooltip();
     const target = screen.getByText("Target Element");
     fireEvent.focus(target);
     expect(await screen.findByText(Message)).toBeInTheDocument();
   });
 
   it("hides tooltip message on blur", async () => {
-    render(<Tooltip Message={Message}>{Children}</Tooltip>);
+    renderTooltip();
     const target = screen.getByText("Target Element");
     fireEvent.focus(target);
     await screen.findByText(Message);
@@ -40,7 +47,7 @@ describe("TooltipArea component", () => {
   });
 
   it("hides tooltip message on pointer leave", async () => {
-    render(<Tooltip Message={Message}>{Children}</Tooltip>);
+    renderTooltip();
     const target = screen.getByText("Target Element");
     fireEvent.pointerEnter(target);
     await screen.findByText(Message);
@@ -50,14 +57,7 @@ describe("TooltipArea component", () => {
 
   it("applies messageElementClassName", async () => {
     const messageElementClassName = "test-message-class";
-    render(
-      <Tooltip
-        Message={Message}
-        messageElementClassName={messageElementClassName}
-      >
-        {Children}
-      </Tooltip>,
-    );
+    renderTooltip({ messageElementClassName });
     const target = screen.getByText("Target Element");
     fireEvent.pointerEnter(target);
     // The message text lives in an inner `.text` span; the class is on the
@@ -69,12 +69,7 @@ describe("TooltipArea component", () => {
   it("does not couple the arrow size to the distance prop", async () => {
     // Arrow size is an independent token; `distance` only drives the gap
     // between the trigger and the popup, not the arrow geometry.
-    const distance = "10px";
-    render(
-      <Tooltip Message={Message} distance={distance}>
-        {Children}
-      </Tooltip>,
-    );
+    renderTooltip({ distance: "10px" });
     const target = screen.getByText("Target Element");
     fireEvent.pointerEnter(target);
     const message = await screen.findByText(Message);
@@ -84,7 +79,7 @@ describe("TooltipArea component", () => {
   });
 
   it("renders the tooltip on the contrasted surface", async () => {
-    render(<Tooltip Message={Message}>{Children}</Tooltip>);
+    renderTooltip();
     const target = screen.getByText("Target Element");
     fireEvent.pointerEnter(target);
     const message = await screen.findByText(Message);
@@ -92,7 +87,7 @@ describe("TooltipArea component", () => {
   });
 
   it("uses createPortal to render tooltip", async () => {
-    render(<Tooltip Message={Message}>{Children}</Tooltip>);
+    renderTooltip();
     const target = screen.getByText("Target Element");
     fireEvent.pointerEnter(target);
     await screen.findByText(Message);
@@ -100,10 +95,15 @@ describe("TooltipArea component", () => {
   });
 
   it("renders tooltip with default parentElement document.body if parentElement not provided", async () => {
-    render(<Tooltip Message={Message}>{Children}</Tooltip>);
+    renderTooltip();
     const target = screen.getByText("Target Element");
     fireEvent.pointerEnter(target);
     await screen.findByText(Message);
     expect(createPortal).toHaveBeenCalledWith(expect.anything(), document.body);
+  });
+
+  it("sets the wrapped component's displayName", () => {
+    const Tooltipped = withTooltip(Trigger, Message);
+    expect(Tooltipped.displayName).toBe("withTooltip(Trigger)");
   });
 });
